+++ carp_core/lib/carp_deployment/carp_core_deployment.g.dart
@@ -58,13 +58,21 @@ DeviceRegistration _$DeviceRegistrationFromJson(Map<String, dynamic> json) {
   )..$type = json[r'$type'] as String?;
 }
 
-Map<String, dynamic> _$DeviceRegistrationToJson(DeviceRegistration instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'registrationCreationDate':
-          instance.registrationCreationDate.toIso8601String(),
-      'deviceId': instance.deviceId,
-    };
+Map<String, dynamic> _$DeviceRegistrationToJson(DeviceRegistration instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['registrationCreationDate'] =
+      instance.registrationCreationDate.toIso8601String();
+  val['deviceId'] = instance.deviceId;
+  return val;
+}
 
 DeviceDeploymentStatus _$DeviceDeploymentStatusFromJson(
     Map<String, dynamic> json) {
@@ -85,11 +93,7 @@ DeviceDeploymentStatus _$DeviceDeploymentStatusFromJson(
 
 Map<String, dynamic> _$DeviceDeploymentStatusToJson(
     DeviceDeploymentStatus instance) {
-  final val = <String, dynamic>{
-    r'$type': instance.$type,
-    'device': instance.device,
-    'requiresDeployment': instance.requiresDeployment,
-  };
+  final val = <String, dynamic>{};
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -97,6 +101,9 @@ Map<String, dynamic> _$DeviceDeploymentStatusToJson(
     }
   }
 
+  writeNotNull(r'$type', instance.$type);
+  val['device'] = instance.device;
+  val['requiresDeployment'] = instance.requiresDeployment;
   writeNotNull('remainingDevicesToRegisterToObtainDeployment',
       instance.remainingDevicesToRegisterToObtainDeployment);
   writeNotNull('remainingDevicesToRegisterBeforeDeployment',
@@ -132,11 +139,7 @@ StudyDeploymentStatus _$StudyDeploymentStatusFromJson(
 
 Map<String, dynamic> _$StudyDeploymentStatusToJson(
     StudyDeploymentStatus instance) {
-  final val = <String, dynamic>{
-    r'$type': instance.$type,
-    'studyDeploymentId': instance.studyDeploymentId,
-    'devicesStatus': instance.devicesStatus,
-  };
+  final val = <String, dynamic>{};
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -144,6 +147,9 @@ Map<String, dynamic> _$StudyDeploymentStatusToJson(
     }
   }
 
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  val['devicesStatus'] = instance.devicesStatus;
   writeNotNull('startTime', instance.startTime?.toIso8601String());
   return val;
 }
@@ -177,11 +183,19 @@ EmailAccountIdentity _$EmailAccountIdentityFromJson(Map<String, dynamic> json) {
 }
 
 Map<String, dynamic> _$EmailAccountIdentityToJson(
-        EmailAccountIdentity instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'emailAddress': instance.emailAddress,
-    };
+    EmailAccountIdentity instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['emailAddress'] = instance.emailAddress;
+  return val;
+}
 
 Participation _$ParticipationFromJson(Map<String, dynamic> json) {
   return Participation(
@@ -263,12 +277,20 @@ DataEndPoint _$DataEndPointFromJson(Map<String, dynamic> json) {
   )..$type = json[r'$type'] as String?;
 }
 
-Map<String, dynamic> _$DataEndPointToJson(DataEndPoint instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'type': instance.type,
-      'dataFormat': instance.dataFormat,
-    };
+Map<String, dynamic> _$DataEndPointToJson(DataEndPoint instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['type'] = instance.type;
+  val['dataFormat'] = instance.dataFormat;
+  return val;
+}
 
 CreateStudyDeployment _$CreateStudyDeploymentFromJson(
     Map<String, dynamic> json) {
@@ -278,11 +300,19 @@ CreateStudyDeployment _$CreateStudyDeploymentFromJson(
 }
 
 Map<String, dynamic> _$CreateStudyDeploymentToJson(
-        CreateStudyDeployment instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'protocol': instance.protocol,
-    };
+    CreateStudyDeployment instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['protocol'] = instance.protocol;
+  return val;
+}
 
 GetStudyDeploymentStatus _$GetStudyDeploymentStatusFromJson(
     Map<String, dynamic> json) {
@@ -292,11 +322,19 @@ GetStudyDeploymentStatus _$GetStudyDeploymentStatusFromJson(
 }
 
 Map<String, dynamic> _$GetStudyDeploymentStatusToJson(
-        GetStudyDeploymentStatus instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-    };
+    GetStudyDeploymentStatus instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  return val;
+}
 
 GetStudyDeploymentStatusList _$GetStudyDeploymentStatusListFromJson(
     Map<String, dynamic> json) {
@@ -310,12 +348,20 @@ GetStudyDeploymentStatusList _$GetStudyDeploymentStatusListFromJson(
 }
 
 Map<String, dynamic> _$GetStudyDeploymentStatusListToJson(
-        GetStudyDeploymentStatusList instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-      'studyDeploymentIds': instance.studyDeploymentIds,
-    };
+    GetStudyDeploymentStatusList instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  val['studyDeploymentIds'] = instance.studyDeploymentIds;
+  return val;
+}
 
 RegisterDevice _$RegisterDeviceFromJson(Map<String, dynamic> json) {
   return RegisterDevice(
@@ -325,13 +371,21 @@ RegisterDevice _$RegisterDeviceFromJson(Map<String, dynamic> json) {
   )..$type = json[r'$type'] as String?;
 }
 
-Map<String, dynamic> _$RegisterDeviceToJson(RegisterDevice instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-      'deviceRoleName': instance.deviceRoleName,
-      'registration': instance.registration,
-    };
+Map<String, dynamic> _$RegisterDeviceToJson(RegisterDevice instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  val['deviceRoleName'] = instance.deviceRoleName;
+  val['registration'] = instance.registration;
+  return val;
+}
 
 UnregisterDevice _$UnregisterDeviceFromJson(Map<String, dynamic> json) {
   return UnregisterDevice(
@@ -340,12 +394,20 @@ UnregisterDevice _$UnregisterDeviceFromJson(Map<String, dynamic> json) {
   )..$type = json[r'$type'] as String?;
 }
 
-Map<String, dynamic> _$UnregisterDeviceToJson(UnregisterDevice instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-      'deviceRoleName': instance.deviceRoleName,
-    };
+Map<String, dynamic> _$UnregisterDeviceToJson(UnregisterDevice instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  val['deviceRoleName'] = instance.deviceRoleName;
+  return val;
+}
 
 GetDeviceDeploymentFor _$GetDeviceDeploymentForFromJson(
     Map<String, dynamic> json) {
@@ -356,12 +418,20 @@ GetDeviceDeploymentFor _$GetDeviceDeploymentForFromJson(
 }
 
 Map<String, dynamic> _$GetDeviceDeploymentForToJson(
-        GetDeviceDeploymentFor instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-      'masterDeviceRoleName': instance.masterDeviceRoleName,
-    };
+    GetDeviceDeploymentFor instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  val['masterDeviceRoleName'] = instance.masterDeviceRoleName;
+  return val;
+}
 
 DeploymentSuccessful _$DeploymentSuccessfulFromJson(Map<String, dynamic> json) {
   return DeploymentSuccessful(
@@ -372,14 +442,22 @@ DeploymentSuccessful _$DeploymentSuccessfulFromJson(Map<String, dynamic> json) {
 }
 
 Map<String, dynamic> _$DeploymentSuccessfulToJson(
-        DeploymentSuccessful instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-      'masterDeviceRoleName': instance.masterDeviceRoleName,
-      'deviceDeploymentLastUpdateDate':
-          instance.deviceDeploymentLastUpdateDate?.toIso8601String(),
-    };
+    DeploymentSuccessful instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  val['masterDeviceRoleName'] = instance.masterDeviceRoleName;
+  val['deviceDeploymentLastUpdateDate'] =
+      instance.deviceDeploymentLastUpdateDate?.toIso8601String();
+  return val;
+}
 
 GetActiveParticipationInvitations _$GetActiveParticipationInvitationsFromJson(
     Map<String, dynamic> json) {
@@ -389,11 +467,19 @@ GetActiveParticipationInvitations _$GetActiveParticipationInvitationsFromJson(
 }
 
 Map<String, dynamic> _$GetActiveParticipationInvitationsToJson(
-        GetActiveParticipationInvitations instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'accountId': instance.accountId,
-    };
+    GetActiveParticipationInvitations instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['accountId'] = instance.accountId;
+  return val;
+}
 
 GetParticipantData _$GetParticipantDataFromJson(Map<String, dynamic> json) {
   return GetParticipantData(
@@ -401,11 +487,19 @@ GetParticipantData _$GetParticipantDataFromJson(Map<String, dynamic> json) {
   )..$type = json[r'$type'] as String?;
 }
 
-Map<String, dynamic> _$GetParticipantDataToJson(GetParticipantData instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-    };
+Map<String, dynamic> _$GetParticipantDataToJson(GetParticipantData instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  return val;
+}
 
 GetParticipantDataList _$GetParticipantDataListFromJson(
     Map<String, dynamic> json) {
@@ -419,12 +513,20 @@ GetParticipantDataList _$GetParticipantDataListFromJson(
 }
 
 Map<String, dynamic> _$GetParticipantDataListToJson(
-        GetParticipantDataList instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-      'studyDeploymentIds': instance.studyDeploymentIds,
-    };
+    GetParticipantDataList instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  val['studyDeploymentIds'] = instance.studyDeploymentIds;
+  return val;
+}
 
 SetParticipantData _$SetParticipantDataFromJson(Map<String, dynamic> json) {
   return SetParticipantData(
@@ -434,10 +536,18 @@ SetParticipantData _$SetParticipantDataFromJson(Map<String, dynamic> json) {
   )..$type = json[r'$type'] as String?;
 }
 
-Map<String, dynamic> _$SetParticipantDataToJson(SetParticipantData instance) =>
-    <String, dynamic>{
-      r'$type': instance.$type,
-      'studyDeploymentId': instance.studyDeploymentId,
-      'inputDataType': instance.inputDataType,
-      'data': instance.data,
-    };
+Map<String, dynamic> _$SetParticipantDataToJson(SetParticipantData instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull(r'$type', instance.$type);
+  val['studyDeploymentId'] = instance.studyDeploymentId;
+  val['inputDataType'] = instance.inputDataType;
+  val['data'] = instance.data;
+  return val;
+}