+++ packages/carp_context_package/lib/src/location/location_datum.dart
@@ -15,38 +15,40 @@ class LocationDatum extends Datum {
 
   LocationDatum() : super();
 
-  LocationDatum.fromLocationDto(LocationDto dto)
-      : latitude = dto.latitude,
-        longitude = dto.longitude,
-        altitude = dto.altitude,
-        accuracy = dto.accuracy,
-        speed = dto.speed,
-        speedAccuracy = dto.speedAccuracy,
-        heading = dto.heading,
-        time = DateTime.fromMillisecondsSinceEpoch(dto.time.toInt()),
-        super();
-
-  // LocationDatum.fromLocation(location.LocationData location)
-  //     : latitude = location.latitude,
-  //       longitude = location.longitude,
-  //       altitude = location.altitude,
-  //       accuracy = location.accuracy,
-  //       speed = location.speed,
-  //       speedAccuracy = location.speedAccuracy,
-  //       heading = location.heading,
-  //       time = DateTime.fromMillisecondsSinceEpoch(location.time.toInt()),
+  // LocationDatum.fromLocationDto(LocationDto dto)
+  //     : latitude = dto.latitude,
+  //       longitude = dto.longitude,
+  //       altitude = dto.altitude,
+  //       accuracy = dto.accuracy,
+  //       speed = dto.speed,
+  //       speedAccuracy = dto.speedAccuracy,
+  //       heading = dto.heading,
+  //       time = DateTime.fromMillisecondsSinceEpoch(dto.time.toInt()),
   //       super();
 
-  LocationDatum.fromPosition(Position position)
-      : latitude = position.latitude,
-        longitude = position.longitude,
-        altitude = position.altitude,
-        accuracy = position.accuracy,
-        speed = position.speed,
-        speedAccuracy = position.speedAccuracy,
-        heading = position.heading,
-        time = position.timestamp,
+  LocationDatum.fromLocation(location.LocationData location)
+      : latitude = location.latitude,
+        longitude = location.longitude,
+        altitude = location.altitude,
+        accuracy = location.accuracy,
+        speed = location.speed,
+        speedAccuracy = location.speedAccuracy,
+        heading = location.heading,
+        time = (location.time != null)
+            ? DateTime.fromMillisecondsSinceEpoch(location.time!.toInt())
+            : null,
         super();
+
+  // LocationDatum.fromPosition(Position position)
+  //     : latitude = position.latitude,
+  //       longitude = position.longitude,
+  //       altitude = position.altitude,
+  //       accuracy = position.accuracy,
+  //       speed = position.speed,
+  //       speedAccuracy = position.speedAccuracy,
+  //       heading = position.heading,
+  //       time = position.timestamp,
+  //       super();
 
   factory LocationDatum.fromJson(Map<String, dynamic> json) =>
       _$LocationDatumFromJson(json);