+++ packages/carp_context_package/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.30.1
+* upgrade to use `location` instead of `carp_background_location`
+* removal of request for local permissions - this has to take place in the app, due to Google's new privacy restrictions.
+
 ## 0.30.0+1
 * upgrade to null-safety
 * update of documentation on `manifest.xml` configuration